# Go_Server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个集合了多种语言的Web靶场，该系统是以食谱菜单管理系统为场景去编写，一种实战化形式的安全漏洞靶场，其中存在多个安全漏洞，需要我们去探索和发现。\\n\\n该项目旨在帮助安全研究人员和爱好者了解和掌握关于不同语言系统的渗透测试和代码审计知识。",
        "title": "食谱菜单管理系统靶场",
        "contact": {
            "name": "Vulnerabilities_Server",
            "url": "https://github.com/A7cc/Vulnerabilities_Server"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/order/add": {
            "post": {
                "tags": [
                    "鉴权接口-订单相关方法"
                ],
                "summary": "新增订单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "添加订单信息",
                        "name": "AddOrderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.AddOrderRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/order/delete/{id}": {
            "delete": {
                "tags": [
                    "鉴权接口-订单相关方法"
                ],
                "summary": "删除订单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "删除订单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/order/detail": {
            "get": {
                "tags": [
                    "鉴权接口-订单相关方法"
                ],
                "summary": "根据ID获取订单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "获取订单ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/order/get": {
            "get": {
                "tags": [
                    "鉴权接口-订单相关方法"
                ],
                "summary": "获取订单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "获取订单查询参数",
                        "name": "GetOrderListRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.GetOrderListRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/role/add": {
            "post": {
                "tags": [
                    "鉴权接口-角色相关方法"
                ],
                "summary": "新增角色信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "添加角色信息",
                        "name": "AddRoleRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.AddRoleRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/role/delete/{id}": {
            "delete": {
                "tags": [
                    "鉴权接口-角色相关方法"
                ],
                "summary": "删除角色信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "删除角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/role/detail": {
            "get": {
                "tags": [
                    "鉴权接口-角色相关方法"
                ],
                "summary": "获取角色详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "获取角色ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/role/get": {
            "get": {
                "tags": [
                    "鉴权接口-角色相关方法"
                ],
                "summary": "获取角色列表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "获取角色查询参数",
                        "name": "GetRoleListRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.GetRoleListRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/role/update": {
            "put": {
                "tags": [
                    "鉴权接口-角色相关方法"
                ],
                "summary": "修改角色信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "修改角色信息参数",
                        "name": "UpdateRoleRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.UpdateRoleRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "define.AddOrderRequest": {
            "type": "object",
            "properties": {
                "food": {
                    "type": "string"
                },
                "num": {
                    "type": "integer"
                },
                "remarks": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "define.AddRoleRequest": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                }
            }
        },
...
        "define.GetOrderListRequest": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "define.GetRoleListRequest": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "define.UpdateRoleRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                }
            }
        },
...
    }
}